Extract verifreg test harness to new module
