+++ actors/verifreg/tests/verifreg_actor_test.rs
@@ -1,41 +1,17 @@
 #![deny(unused_must_use)] // Force unwrapping Result<_, Err>
 
-use fvm_shared::{HAMT_BIT_WIDTH, MethodNum};
-use fvm_shared::address::Address;
-use fvm_shared::bigint::bigint_ser::BigIntDe;
-use fvm_shared::blockstore::MemoryBlockstore;
-use fvm_shared::encoding::RawBytes;
 use lazy_static::lazy_static;
 
 use fil_actor_verifreg::{
-    Actor as VerifregActor, AddVerifierClientParams, AddVerifierParams, DataCap, Method, MINIMUM_VERIFIED_DEAL_SIZE,
-    State,
+    DataCap,
+    MINIMUM_VERIFIED_DEAL_SIZE,
 };
-use fil_actors_runtime::{
-    ActorError, make_empty_map, make_map_with_root_and_bitwidth, Map, SYSTEM_ACTOR_ADDR,
-};
-use fil_actors_runtime::test_utils::*;
+
+mod harness;
 
 lazy_static! {
-    static ref ROOT_ADDR: Address = Address::new_id(101);
     static ref VERIFIER_ALLOWANCE: DataCap = MINIMUM_VERIFIED_DEAL_SIZE.clone() + DataCap::from(42);
     static ref CLIENT_ALLOWANCE: DataCap = VERIFIER_ALLOWANCE.clone() - DataCap::from(1);
-}
-
-fn construct_runtime() -> MockRuntime {
-    MockRuntime {
-        receiver: *ROOT_ADDR,
-        caller: *SYSTEM_ACTOR_ADDR,
-        caller_type: *SYSTEM_ACTOR_CODE_ID,
-        ..Default::default()
-    }
-}
-
-fn make_harness() -> (Harness, MockRuntime) {
-    let mut rt = construct_runtime();
-    let h = Harness { root: *ROOT_ADDR };
-    h.construct_and_verify(&mut rt, &h.root);
-    return (h, rt);
 }
 
 mod construction {
@@ -48,11 +24,12 @@ mod construction {
     use fil_actors_runtime::SYSTEM_ACTOR_ADDR;
     use fil_actors_runtime::test_utils::*;
 
-    use crate::{construct_runtime, Harness};
+    use crate::harness;
+    use harness::*;
 
     #[test]
     fn construct_with_root_id() {
-        let mut rt = construct_runtime();
+        let mut rt = new_runtime();
         let h = Harness { root: Address::new_id(101) };
         h.construct_and_verify(&mut rt, &h.root);
         h.check_state();
@@ -60,7 +37,7 @@ mod construction {
 
     #[test]
     fn construct_resolves_non_id() {
-        let mut rt = construct_runtime();
+        let mut rt = new_runtime();
         let h = Harness { root: Address::new_id(101) };
         let root_pubkey = Address::new_bls(&[7u8; BLS_PUB_LEN]).unwrap();
         rt.id_addresses.insert(root_pubkey, h.root);
@@ -70,7 +47,7 @@ mod construction {
 
     #[test]
     fn construct_fails_if_root_unresolved() {
-        let mut rt = construct_runtime();
+        let mut rt = new_runtime();
         let root_pubkey = Address::new_bls(&[7u8; BLS_PUB_LEN]).unwrap();
 
         rt.expect_validate_caller_addr(vec![*SYSTEM_ACTOR_ADDR]);
@@ -96,12 +73,12 @@ mod verifiers {
     };
     use fil_actors_runtime::test_utils::*;
 
-    use crate::{make_harness, ROOT_ADDR, VERIFIER_ALLOWANCE};
+    use crate::{harness, VERIFIER_ALLOWANCE};
+    use harness::*;
 
     #[test]
     fn add_verifier_requires_root_caller() {
-        let (h, mut rt) = make_harness();
-
+        let (h, mut rt) = new_harness();
         rt.expect_validate_caller_addr(vec![h.root]);
         rt.set_caller(*VERIFREG_ACTOR_CODE_ID, Address::new_id(501));
         let params = AddVerifierParams {
@@ -120,7 +97,7 @@ mod verifiers {
 
     #[test]
     fn add_verifier_enforces_min_size() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let allowance = MINIMUM_VERIFIED_DEAL_SIZE.clone() - DataCap::from(1);
         expect_abort(
             ExitCode::ErrIllegalArgument,
@@ -131,7 +108,7 @@ mod verifiers {
 
     #[test]
     fn add_verifier_rejects_root() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         expect_abort(
             ExitCode::ErrIllegalArgument,
             h.add_verifier(&mut rt, &ROOT_ADDR, &VERIFIER_ALLOWANCE),
@@ -141,7 +118,7 @@ mod verifiers {
 
     #[test]
     fn add_verifier_rejects_client() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let client = Address::new_id(202);
         h.add_verifier_and_client(
             &mut rt,
@@ -159,7 +136,7 @@ mod verifiers {
 
     #[test]
     fn add_verifier_rejects_unresolved_address() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier_key_address = Address::new_secp256k1(&[3u8; 65]).unwrap();
         // Expect runtime to attempt to create the actor, but don't add it to the mock's
         // address resolution table.
@@ -180,14 +157,14 @@ mod verifiers {
 
     #[test]
     fn add_verifier_id_address() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         h.add_verifier(&mut rt, &Address::new_id(201), &VERIFIER_ALLOWANCE).unwrap();
         h.check_state();
     }
 
     #[test]
     fn add_verifier_resolves_address() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let pubkey_addr = Address::new_secp256k1(&[0u8; 65]).unwrap();
         rt.id_addresses.insert(pubkey_addr, Address::new_id(201));
 
@@ -197,7 +174,7 @@ mod verifiers {
 
     #[test]
     fn remove_requires_root() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
 
@@ -215,7 +192,7 @@ mod verifiers {
 
     #[test]
     fn remove_requires_verifier_exists() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(501);
         expect_abort(ExitCode::ErrIllegalArgument, h.remove_verifier(&mut rt, &verifier));
         h.check_state();
@@ -223,7 +200,7 @@ mod verifiers {
 
     #[test]
     fn remove_verifier() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
         h.remove_verifier(&mut rt, &verifier).unwrap();
@@ -232,7 +209,7 @@ mod verifiers {
 
     #[test]
     fn remove_verifier_id_address() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier_pubkey = Address::new_bls(&[1u8; BLS_PUB_LEN]).unwrap();
         let verifier_id = Address::new_id(201);
         rt.id_addresses.insert(verifier_pubkey, verifier_id);
@@ -257,11 +234,12 @@ mod clients {
     };
     use fil_actors_runtime::test_utils::*;
 
-    use crate::{CLIENT_ALLOWANCE, make_harness, VERIFIER_ALLOWANCE};
+    use crate::{CLIENT_ALLOWANCE, harness, VERIFIER_ALLOWANCE};
+    use harness::*;
 
     #[test]
     fn many_verifiers_and_clients() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier1 = Address::new_id(201);
         let verifier2 = Address::new_id(202);
 
@@ -292,7 +270,7 @@ mod clients {
 
     #[test]
     fn verifier_allowance_exhausted() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         // Verifier only has allowance for one client.
         h.add_verifier(&mut rt, &verifier, &CLIENT_ALLOWANCE).unwrap();
@@ -313,7 +291,7 @@ mod clients {
 
     #[test]
     fn resolves_client_address() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
 
         let client_pubkey = Address::new_bls(&[7u8; BLS_PUB_LEN]).unwrap();
         let client_id = Address::new_id(301);
@@ -335,7 +313,7 @@ mod clients {
 
     #[test]
     fn minimum_allowance_ok() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
 
@@ -347,7 +325,7 @@ mod clients {
 
     #[test]
     fn rejects_unresolved_address() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
 
@@ -372,7 +350,7 @@ mod clients {
 
     #[test]
     fn rejects_allowance_below_minimum() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
 
@@ -387,7 +365,7 @@ mod clients {
 
     #[test]
     fn rejects_non_verifier_caller() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
 
@@ -409,7 +387,7 @@ mod clients {
 
     #[test]
     fn rejects_allowance_greater_than_verifier_cap() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
 
@@ -423,7 +401,7 @@ mod clients {
 
     #[test]
     fn rejects_root_as_client() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
         expect_abort(
@@ -435,7 +413,7 @@ mod clients {
 
     #[test]
     fn rejects_verifier_as_client() {
-        let (h, mut rt) = make_harness();
+        let (h, mut rt) = new_harness();
         let verifier = Address::new_id(201);
         h.add_verifier(&mut rt, &verifier, &VERIFIER_ALLOWANCE).unwrap();
         expect_abort(
@@ -458,151 +436,4 @@ mod clients {
 
         h.check_state();
     }
-}
-
-///// Test harness /////
-
-struct Harness {
-    root: Address,
-}
-
-impl Harness {
-    fn construct_and_verify(&self, rt: &mut MockRuntime, root_param: &Address) {
-        rt.expect_validate_caller_addr(vec![*SYSTEM_ACTOR_ADDR]);
-        let ret = rt
-            .call::<VerifregActor>(
-                Method::Constructor as MethodNum,
-                &RawBytes::serialize(root_param).unwrap(),
-            )
-            .unwrap();
-
-        assert_eq!(RawBytes::default(), ret);
-        rt.verify();
-
-        let empty_map = make_empty_map::<_, ()>(&rt.store, HAMT_BIT_WIDTH).flush().unwrap();
-
-        let state: State = rt.get_state().unwrap();
-
-        assert_eq!(self.root, state.root_key);
-        assert_eq!(empty_map, state.verified_clients);
-        assert_eq!(empty_map, state.verifiers);
-    }
-
-    fn add_verifier(
-        &self,
-        rt: &mut MockRuntime,
-        verifier: &Address,
-        allowance: &DataCap,
-    ) -> Result<(), ActorError> {
-        rt.expect_validate_caller_addr(vec![self.root]);
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, self.root);
-        let params = AddVerifierParams { address: *verifier, allowance: allowance.clone() };
-        let ret = rt.call::<VerifregActor>(
-            Method::AddVerifier as MethodNum,
-            &RawBytes::serialize(params).unwrap(),
-        )?;
-        assert_eq!(RawBytes::default(), ret);
-        rt.verify();
-
-        self.assert_verifier_allowance(rt, &verifier, allowance);
-        Ok(())
-    }
-
-    fn remove_verifier(&self, rt: &mut MockRuntime, verifier: &Address) -> Result<(), ActorError> {
-        rt.expect_validate_caller_addr(vec![self.root]);
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, self.root);
-        let ret = rt.call::<VerifregActor>(
-            Method::RemoveVerifier as MethodNum,
-            &RawBytes::serialize(verifier).unwrap(),
-        )?;
-        assert_eq!(RawBytes::default(), ret);
-        rt.verify();
-
-        self.assert_verifier_removed(rt, verifier);
-        Ok(())
-    }
-
-    fn assert_verifier_allowance(&self, rt: &MockRuntime, verifier: &Address, allowance: &DataCap) {
-        let verifier_id_addr = rt.get_id_address(verifier).unwrap();
-        assert_eq!(*allowance, self.get_verifier_allowance(rt, &verifier_id_addr));
-    }
-
-    fn get_verifier_allowance(&self, rt: &MockRuntime, verifier: &Address) -> DataCap {
-        let verifiers = load_verifiers(&rt);
-        let BigIntDe(allowance) = verifiers.get(&verifier.to_bytes()).unwrap().unwrap();
-        return allowance.clone();
-    }
-
-    fn assert_verifier_removed(&self, rt: &MockRuntime, verifier: &Address) {
-        let verifier_id_addr = rt.get_id_address(verifier).unwrap();
-        let verifiers = load_verifiers(&rt);
-        assert_eq!(false, verifiers.contains_key(&verifier_id_addr.to_bytes()).unwrap())
-    }
-
-    fn add_client(
-        &self,
-        rt: &mut MockRuntime,
-        verifier: &Address,
-        client: &Address,
-        allowance: &DataCap,
-        expected_allowance: &DataCap,
-    ) -> Result<(), ActorError> {
-        rt.expect_validate_caller_any();
-        rt.set_caller(*ACCOUNT_ACTOR_CODE_ID, *verifier);
-        let params = AddVerifierClientParams { address: *client, allowance: allowance.clone() };
-        let ret = rt.call::<VerifregActor>(
-            Method::AddVerifiedClient as MethodNum,
-            &RawBytes::serialize(params).unwrap(),
-        )?;
-        assert_eq!(RawBytes::default(), ret);
-        rt.verify();
-
-        // Confirm the verifier was added to state.
-        let client_id_addr = rt.get_id_address(&client).unwrap();
-        assert_eq!(*expected_allowance, self.get_client_allowance(rt, &client_id_addr));
-        Result::Ok(())
-    }
-
-    fn assert_client_allowance(&self, rt: &MockRuntime, client: &Address, allowance: &DataCap) {
-        let client_id_addr = rt.get_id_address(client).unwrap();
-        assert_eq!(*allowance, self.get_client_allowance(rt, &client_id_addr));
-    }
-
-    fn get_client_allowance(&self, rt: &MockRuntime, client: &Address) -> DataCap {
-        let clients = load_clients(&rt);
-        let BigIntDe(allowance) = clients.get(&client.to_bytes()).unwrap().unwrap();
-        return allowance.clone();
-    }
-
-    fn add_verifier_and_client(
-        &self,
-        rt: &mut MockRuntime,
-        verifier: &Address,
-        client: &Address,
-        verifier_allowance: &DataCap,
-        client_allowance: &DataCap,
-    ) {
-        self.add_verifier(rt, verifier, verifier_allowance).unwrap();
-        self.add_client(rt, verifier, client, client_allowance, client_allowance).unwrap();
-    }
-
-    fn check_state(&self) {
-        // TODO: https://github.com/filecoin-project/builtin-actors/issues/44
-    }
-}
-
-fn load_verifiers(rt: &MockRuntime) -> Map<MemoryBlockstore, BigIntDe> {
-    let state: State = rt.get_state().unwrap();
-    make_map_with_root_and_bitwidth::<_, BigIntDe>(&state.verifiers, &rt.store, HAMT_BIT_WIDTH)
-        .unwrap()
-}
-
-fn load_clients(rt: &MockRuntime) -> Map<MemoryBlockstore, BigIntDe> {
-    let state: State = rt.get_state().unwrap();
-    make_map_with_root_and_bitwidth::<_, BigIntDe>(
-        &state.verified_clients,
-        &rt.store,
-        HAMT_BIT_WIDTH,
-    )
-    .unwrap()
 }
